requirements.txt
requests
pandas
openpyxl
python-dateutil==2.8.2
tqdm